pyproject.toml
[tool.poetry]
name = "airouter"
version = "0.1.0"
description = ""
authors = ["Laurentiu Piciu <laurentiupiciu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
openai = "^1.2.0"
boto3 = "^1.28.82"
botocore = "^1.31.82"
pydantic = "^2.4.2"
tenacity = "^8.0"
google-cloud-aiplatform = "^1.36.1"
python-decouple = "^3.8"
tiktoken = "^0.5.1"
urwid = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
